Chore: Basic setup to start a recipe activity from RecipeBookFragment and some UI changes
