+++ app/src/main/AndroidManifest.xml
@@ -13,6 +13,13 @@
         android:theme="@style/Theme.WellFed"
         tools:targetApi="31">
         <activity
+            android:name=".Recipe.RecipeActivity"
+            android:exported="false">
+            <meta-data
+                android:name="android.app.lib_name"
+                android:value="" />
+        </activity>
+        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
