+++ app/src/main/java/com/example/wellfed/MainActivity.java
@@ -1,5 +1,6 @@
 package com.example.wellfed;
 
+import androidx.appcompat.app.AppCompatDelegate;
 import androidx.fragment.app.FragmentActivity;
 import androidx.viewpager2.widget.ViewPager2;
