+++ app/src/main/java/com/example/wellfed/NavigationCollectionAdapter.java
@@ -5,6 +5,8 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentActivity;
 import androidx.viewpager2.adapter.FragmentStateAdapter;
 
+import com.example.wellfed.Recipe.RecipeBookFragment;
+
 public class NavigationCollectionAdapter extends FragmentStateAdapter {
     public NavigationCollectionAdapter(FragmentActivity fa) {
         super(fa);
