+++ app/src/main/java/com/example/wellfed/Recipe/RecipeActivity.java
@@ -1,0 +1,16 @@
+package com.example.wellfed.Recipe;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.example.wellfed.R;
+
+public class RecipeActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_recipe);
+    }
+}
