+++ app/src/main/res/layout/activity_recipe.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".Recipe.RecipeActivity"
+    android:orientation="vertical">
+    <ImageView
+        android:id="@+id/recipe_img"
+        android:layout_width="match_parent"
+        android:layout_height="200dp"
+        android:src="@drawable/taco"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        android:scaleType="centerCrop"/>
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/recipe_title_textView"
+        app:layout_constraintTop_toBottomOf="@+id/recipe_img"
+        app:layout_constraintStart_toStartOf="parent"
+        android:text="Tacos"
+        android:textSize="22sp"
+        android:layout_marginLeft="16dp"
+        android:textStyle="bold"/>
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/recipe_prep_time_textView"
+        app:layout_constraintTop_toBottomOf="@+id/recipe_title_textView"
+        app:layout_constraintStart_toStartOf="parent"
+        android:layout_marginLeft="16dp"
+        android:text="Preparation Time: 20 minutes"/>
+
+</androidx.constraintlayout.widget.ConstraintLayout>
